+++ scripts/aliases.ps1
@@ -1,0 +1,4 @@
+### © Copyright 2024 Frankie Homewood <F.Homewood@outlook.com>
+# An alias script to group all aliases into one file to be sourced.
+
+Set-Alias -Name "activate" -Value "Activate-VEnv"