+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Meeting - A command to create templated meeting notes.
+- Meeting - Keeping in Touch template file.
 
 
 ## [0.1.2] - 2024-01-18