+++ CHANGELOG.md
@@ -5,8 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
-- Functions script groups all shared functions in to one file to be sourced.
-- Aliases script groups all in to one file to be sourced.
+- Functions script groups all shared functions into one file to be sourced.
+- Aliases script groups all into one file to be sourced.
 - `activate` alias to be used to jump into a python virtual environment quickly.