+++ CHANGELOG.md
@@ -5,12 +5,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
-- Powershell Profile script.
+- PowerShell Profile script.
 - Functions script groups all shared functions into one file to be sourced in profile.
 - Aliases script groups all into one file to be sourced in profile.
 - Prompt script stores the prompt into one space to be sourced whenever the prompt needs overriding.
 - `activate` alias to be used to jump into a python virtual environment quickly.
-- `profile` alias reloads the current profile in the current powershell.
+- `profile` alias reloads the current profile in the current PowerShell.
 
 
 ## [0.1.2] - 2024-01-18